+++ packages/Testing/src/TestCases/UnitTestCase.php
@@ -32,6 +32,9 @@ abstract class UnitTestCase extends Unit
      */
     protected function _before()
     {
+//        error_reporting(-1);
+//        ini_set('display_errors', true);
+
         $this->setupFaker();
     }